% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{busiest_routes}
\alias{busiest_routes}
\title{Find busiest air routes.}
\usage{
busiest_routes(dataframe, origcol, destcol)
}
\arguments{
\item{dataframe}{data with flight records (disaggregate)}

\item{origcol}{column identifying origin airport or city}

\item{destcol}{column identifying destination city or airport}
}
\value{
sorted list of busiest routes
}
\description{
This finds the busiest (bidirectional) air routes in the country.
}
