% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load data from a data file, and supplemental files for city and carrier.}
\usage{
load_data(datafile, cityfile, carrierfile)
}
\arguments{
\item{datafile}{data to load}

\item{cityfile}{City ID to name mapping}

\item{carrierfile}{Carrier ID to name mapping}
}
\value{
loaded and joined data
}
\description{
Load data from a data file, and supplemental files for city and carrier.
}
