% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{market_shares}
\alias{market_shares}
\title{Find market shares.}
\usage{
market_shares(dataframe, origcol, carriercol)
}
\arguments{
\item{dataframe}{data with airline tickets}

\item{origcol}{origin city or airport}

\item{carriercol}{air carrier}
}
\value{
data frame of market shares by airport and carrier
}
\description{
Find market shares.
}
